+++ src/quakestats/dataprovider/__init__.py
@@ -2,12 +2,8 @@ from quakestats.dataprovider.feeder import (
     FeedFull,
     MatchFeeder,
 )
-from quakestats.dataprovider.match import (
-    FullMatchInfo,
-)
 
 __all__ = [
     "FeedFull",
     "MatchFeeder",
-    "FullMatchInfo",
 ]
